[FIX] Refactor moving fetch data to folder services
